Busca de produtos por nome e categorias

+++ src/main/java/com/digital/cursomc/dto/ClienteNewDTO.java
@@ -53,7 +53,7 @@ public class ClienteNewDTO implements Serializable {
 	
 	private String telefone3;
 
-	private Long cidadeId;
+	private Integer cidadeId;
 	
 	 
 }
