+++ src/main/java/com/digital/cursomc/repositories/CategoriaRepository.java
@@ -1,5 +1,6 @@
 package com.digital.cursomc.repositories;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +9,9 @@ import org.springframework.stereotype.Repository;
 import com.digital.cursomc.domain.Categoria;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+
+//	List<Categoria> findAll(List<Integer> ids);
  
 //	Categoria findById(long id);
 }
